init tyck (#308)
* init tyck

* well, fix for now?

* ???

* well

* [autofix.ci] apply automated fixes

* well

* .

* try

* [autofix.ci] apply automated fixes

* Add relplot dependency to pyproject.toml

* Update ignore-missing-imports in pyrefly.toml

Remove 'relplot' from the ignore-missing-imports list.

* Update autofix.yml

* Remove version from setup-uv action in autofix.yml

Removed version specification for setup-uv action.

* [autofix.ci] apply automated fixes

---------

+++ rwkv/model/rwkv_model.py
@@ -33,7 +33,7 @@ class RWKV7Config:
     )
     n_heads: int
     n_layers: int
-    channel_mixer_factor: int
+    channel_mixer_factor: float
 
     # For stacking RWKV7 on top of one-another. We allow sending in the total number of layers and a layer offset so that we can achieve better initialization
     layer_offset: int
